+++ src/api/l_collider.c
@@ -33,10 +33,10 @@ static int l_lovrColliderGetShapes(lua_State* L) {
   Collider* collider = luax_checktype(L, 1, Collider);
   size_t count;
   Shape** shapes = lovrColliderGetShapes(collider, &count);
-  lua_createtable(L, count, 0);
+  lua_createtable(L, (int) count, 0);
   for (size_t i = 0; i < count; i++) {
     luax_pushshape(L, shapes[i]);
-    lua_rawseti(L, -2, i + 1);
+    lua_rawseti(L, -2, (int) i + 1);
   }
   return 1;
 }
@@ -45,10 +45,10 @@ static int l_lovrColliderGetJoints(lua_State* L) {
   Collider* collider = luax_checktype(L, 1, Collider);
   size_t count;
   Joint** joints = lovrColliderGetJoints(collider, &count);
-  lua_createtable(L, count, 0);
+  lua_createtable(L, (int) count, 0);
   for (size_t i = 0; i < count; i++) {
     luax_pushjoint(L, joints[i]);
-    lua_rawseti(L, -2, i + 1);
+    lua_rawseti(L, -2, (int) i + 1);
   }
   return 1;
 }
